backend/pyproject.toml
[tool.poetry]
name = "glp1-coach-backend"
version = "1.0.0"
description = "GLP-1 Coach Backend API and Worker"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
supabase = "^2.0.0"
anthropic = "^0.15.0"
langfuse = "^2.20.0"
sentry-sdk = "^1.38.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
httpx = "^0.25.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']